fix: Correct datatype for Exemplar value uint64->int64 (#4779)
* fix: Correct datatype for Exemplar value uint64->int64

* Fix missing test

+++ pkg/model/heatmap/range.go
@@ -336,7 +336,7 @@ func pointToExemplar(
 		Timestamp: point.Timestamp,
 		ProfileId: profileID,
 		SpanId:    spanIDStr,
-		Value:     uint64(point.Value),
+		Value:     point.Value,
 		Labels:    filteredLabels,
 	}
 }
